Crate stm

This library implements software transactional memory, often abbreviated with STM.

It is designed closely to haskells STM library. Read Simon Marlow's Parallel and Concurrent Programming in Haskell for more info. Especially the chapter about Performance is also important for using STM in rust.

With locks the sequence of two threadsafe actions is no longer threadsafe because other threads may interfer in between of these actions. Applying another lock may lead to common sources of errors like deadlocks and forgotten locks.

Unlike locks Software transactional memory is composable. It is typically implemented by writing all read and write operations in a log. When the action has finished, the reads are checked for consistency and depending on the result either the writes are committed in a single atomic operation or the result is discarded and the computation run again.

Usage

You should only use the functions that are safe to use.

Don't have side effects except for the atomic variables, from this library. Especially a mutex or other blocking mechanisms inside of software transactional memory is dangerous.

You can run the top-level atomic operation by calling atomically.

use stm::atomically;
atomically(|trans| {
    // some action
    // return value as `Result`, for example
    Ok(42)
});

Calls to atomically should not be nested.

For running an atomic operation inside of another, pass a mutable reference to a Transaction and call try! on the result. You should not handle the error yourself.

use stm::{atomically, TVar};
let var = TVar::new(0);

let x = atomically(|trans| {
    try!(var.write(trans, 42));
    var.read(trans) // return the value saved in var
});

println!("var = {}", x);

STM safety

Software transactional memory is completely safe in the terms, that rust considers safe. Still there are multiple rules that you should obey when dealing with software transactional memory:

Don't run code with side effects, especially no IO-code, because stm is designed to be run multiple times. Return a closure if you have to.

Don't handle the error types yourself, unless you absolutely know, what you are doing. Use Transaction::or, to try alternatives.

Don't run atomically inside of another, because your thread will immediately panic. When you use STM in the inner of a function then express that in the public interface, by taking a &mut Transaction and returning a StmResult<T>. Callers can safely compose it into larger blocks.

Don't mix locks and transactions. Your code will easily deadlock and slow down on unpredictably.

When you put an Arc into a Var, don't use inner mutability to modify it because, the inner still points to the original value.

Speed

Generally keep your atomic blocks as small as possible, because the more time you spend, the more likely it is, to collide with other threads. For STM, reading vars is quite slow, because it needs to look them up in the log every time, they are written to, and every used var increases the chance of collisions. You should keep the amount of accessed variables as low as needed.

Structs

TVar
A variable that can be used in a STM-Block

Transaction
Transaction tracks all the read and written variables.

Enums

StmError

Functions

atomically
Run a function atomically by using Software Transactional Memory. It calls to Transaction::with internally, but is more explicit.

retry
call retry, to abort an operation. It takes another path of an Transaction::or or blocks until any variable changes.

Type Definitions

StmResult
StmResult is a result of a single step of a STM calculation.