Enum stm::StmResult

pub enum StmResult<T> {
    Success(T),
    Failure,
    Retry,
}

a result of each step of a STM calculation

it informs of the success or the type of failure

Variants

Success

the call succeeded

Failure

the call failed. immediate return

is used when a variable is no longer consistent

Retry

retry was called

may block until at least one read variable has changed