Struct stm::STM

pub struct STM<T> {
    // some fields omitted
}

class representing a STM computation

Methods

impl<T: 'static> STM<T>

fn new<F>(func: F) -> STM<T> where F: Fn() -> StmResult<T> + 'static

create a new STM calculation from a closure

fn intern_run(&self) -> StmResult<T>

run a computation and return the result

internal use only. Prefer atomically because it sets up the log and retry the computation until it has succeeded

internal use only

fn atomically(&self) -> T

run a STM computation atomically

fn or(self, other: STM<T>) -> STM<T>

if one of both computations fails with a call to retry then run the other one

if both call retry then the thread will block until any of the vars that were read in one of the both branches changes

fn and<R: 'static>(self, other: STM<R>) -> STM<R>

run the first and afterwards the second one

first.and(second) is equal to

stm!({
    stm_call!(first);
    stm_call!(second)
});

fn and_then<F, R>(self, f: F) -> STM<R> where F: Fn(T) -> STM<R>, F: 'static, R: 'static

run the first and then applies the return value to the function f which returns a STM-Block that is then executed

first.and_then(second) is equal to

stm!({
    let x = stm_call!(first);
    stm_call!(second(x))
});