Struct stm::TVar

pub struct TVar<T> {
    // some fields omitted
}

A variable that can be used in a STM-Block

Methods

impl<T> TVar<T> where T: Any + Sync + Send + Clone

fn new(val: T) -> TVar<T>

create a new var

fn read_atomic(&self) -> T

read a value atomically

this should be called from outside of stm and is faster than wrapping a read in STM but is not composable

read_atomic returns a clone of the value.

If the value contains a shared reference mutating it is a side effect which may break STM-semantics

This is a faster alternative to

use stm::*;

let var = TVar::new(0);
atomically(|trans| var.read(trans));

fn read_ref_atomic(&self) -> Arc<Any + Send + Sync>

read a value atomically but return a reference

this is mostly used internally but can be useful in certain cases where the additional clone performed by read_atomic is unwanted

fn read(&self, transaction: &mut Transaction) -> StmResult<T>

The normal way to access a var.

It is equivalent to transaction.read(&var), but more ergonomic.

fn write(&self, transaction: &mut Transaction, value: T) -> StmResult<()>

The normal way to write a var.

It is equivalent to transaction.write(&var, value), but more ergonomic.

fn control_block(&self) -> &Arc<VarControlBlock>

access the control block of the var

internal use only