Struct stm::Transaction

pub struct Transaction {
    // some fields omitted
}

Transaction tracks all the read and written variables.

It is used for checking vars, to ensure atomicity.

Methods

impl Transaction

fn with<T, F>(f: F) -> T where F: Fn(&mut Transaction) -> StmResult<T>

Run a function with a transaction.

It is equivalent to atomically.

fn read<T: Send + Sync + Any + Clone>(&mut self, var: &TVar<T>) -> StmResult<T>

Read a variable and return the value.

this is not always consistent with the current value of the var but may be an outdated or written but not commited value

The used code should be capable of handling inconsistens states without running into infinite loops. Just the commit of wrong values is prevented.

fn write<T: Any + Send + Sync + Clone>(&mut self, var: &TVar<T>, value: T) -> StmResult<()>

Write a variable.

write does not immediately change the value, but atomically commits all writes at the end of the computation.

fn or<T, F1, F2>(&mut self, first: F1, second: F2) -> StmResult<T> where F1: Fn(&mut Transaction) -> StmResult<T>, F2: Fn(&mut Transaction) -> StmResult<T>

Combine two calculations. When one blocks with retry, run the other, but don't commit the changes in the first. It still blocks on all, so that you can wait for one of two alternatives.